Refactor: move get-field-title and sort-fields to customNodes folder
