+++ src/frontend/src/customNodes/utils/get-field-title.tsx
@@ -1,0 +1,10 @@
+import { APITemplateType } from "../../types/api";
+
+export default function getFieldTitle(
+    template: APITemplateType,
+    templateField: string,
+  ): string {
+    return template[templateField].display_name
+      ? template[templateField].display_name!
+      : template[templateField].name ?? templateField;
+  }
